move patch_motors to dodal.testing
