+++ src/dodal/testing/__init__.py
@@ -1,0 +1,3 @@
+from .setup import patch_all_motors, patch_motor
+
+__all__ = ["patch_motor", "patch_all_motors"]
