+++ tests/devices/test_gridscan.py
@@ -22,7 +22,7 @@ from dodal.devices.fast_grid_scan import (
     set_fast_grid_scan_params,
 )
 from dodal.devices.smargon import Smargon
-from dodal.devices.util.test_utils import patch_all_motors
+from dodal.testing import patch_all_motors
 
 
 def discard_status(st: Status | DeviceStatus):
