+++ tests/devices/test_motors.py
@@ -12,7 +12,7 @@ from dodal.devices.motors import (
     XYZPitchYawRollStage,
     XYZThetaStage,
 )
-from dodal.devices.util.test_utils import patch_all_motors
+from dodal.testing import patch_all_motors
 
 
 @pytest.fixture
