+++ tests/devices/test_smargon.py
@@ -8,7 +8,7 @@ from ophyd_async.core import init_devices, observe_value
 from ophyd_async.testing import get_mock_put, set_mock_value
 
 from dodal.devices.smargon import CombinedMove, DeferMoves, Smargon, StubPosition
-from dodal.devices.util.test_utils import patch_all_motors
+from dodal.testing import patch_all_motors
 
 
 @pytest.fixture
